+++ CHANGELOG.md
@@ -5,3 +5,6 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/eps1lon/poe-components-item/compare/v0.0.1...HEAD)
+# Changed
+* Popups can  now have either a fixed width or compute their width which 
+  should result in the same width as items on the PoE website.